pyproject.toml
[tool.poetry]
name = "infra-tools"
version = "0.1.0"
description = ""
authors = ["MU-Software <musoftware@mudev.cc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
infralib = {path = "infralib", develop = true}
pydantic = "^2.4.2"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"
setuptools = "^68.2.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
